! Sync to SMF 10497  ! r10489 Setting base_theme_dir caused undefined index errors in manage languages area. (ManageServer.php) [Bug 4634]  ! r10490 Personal message preview did not display post errors. (PersonalMessages.php) [Bug 4653]  ! r10491-2 Custom language strings were not loaded when editing theme settings. (Themes.php) [Bug 4658]  ! r10493 modify_modifications hook should be triggered earlier in the code. (ManageSettings.php) [Bug 4638]  ! r10494-5 menu_items hook was triggered too late in the code. (Subs.php) [Bug 4637,4670]  ! r10496 (adapted) Undefined index errors when there were not any recent posts. (BoardIndexInfoCenter.template.php) [Bug 4650]  ! r10497 CAPTCHA settings were linked to the incorrect area in admin search results. (Admin.php) [Bug 4707]
I'm not entirely happy with the changes in r10494/5, as per my comments originally in bug 4637; I do not entirely agree that the menu items hooked items should be cached like this unless you guarantee to flush the cache at any point when you will removing add-ons, or you have add-ons that do funky dynamic buttons. (Any add-ons should really be paying attention. It isn't that hard, either. But I can see the flip side of why it's been done, too, that it does simplify and make consistent additive item behaviour. So I've committed it, after some internal deliberation. It will be interesting to see how this stands up for add-ons that actually do this; I suspect no-one's actually been using the hooks for this other than Fel (for PortaMX), Live (for EP) and myself (SimpleDesk) anyway. Won't get into the debate here over whether a system author should listen to the needs of those who modify the system or not.)