! The revised admin panel (lots of files, including new ones and getting rid of old ones)
The main thing that's different is the front page, and there's still some compatibility code in there from earlier drafts, this isn't finished at all - more icons are needed, as well as completing the move of some features. In the end, Core Features should ultimately be removed from the system, as the first step.