Fix low spam/flood limits
spamProtection() did not handle search limits correct.
It was not possible to disable flood control for search,
lowest possible value was 1. The new check also allows
a value of 0.
If user has moderator permissions, the limit was
always overridden with 2. We now check if the limit is
greater than 2 and if not, we override it with 2.